fix/댓글 조회 에러, 식당 전체 조회 에러, 회원 수정에러, 좋아요 삭제 에러 등 코드 보완

+++ server/src/main/java/com/main/project/thumbUp/service/ThumbUpServiceImpl.java
@@ -1,10 +1,13 @@
 package com.main.project.thumbUp.service;
 
+import com.main.project.exception.BusinessLogicException;
+import com.main.project.exception.ExceptionCode;
 import com.main.project.review.entity.Review;
 import com.main.project.review.repository.ReviewRepository;
 import com.main.project.thumbUp.entity.ThumbUp;
 import com.main.project.thumbUp.repository.ThumbUpRepository;
 import com.main.project.user.entity.WebUser;
+import com.main.project.user.repository.UserRepository;
 import com.main.project.user.service.UserService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -24,12 +27,13 @@ public class ThumbUpServiceImpl implements ThumbUpService{
 
     private final ThumbUpRepository thumbUpRepository;
     ReviewRepository reviewRepository;
-    UserService userService;
+    UserRepository userRepository;
 
-    public ThumbUpServiceImpl(ThumbUpRepository thumbUpRepository, ReviewRepository reviewRepository, UserService userService) {
+
+    public ThumbUpServiceImpl(ThumbUpRepository thumbUpRepository, ReviewRepository reviewRepository, UserRepository userRepository) {
         this.thumbUpRepository = thumbUpRepository;
         this.reviewRepository = reviewRepository;
-        this.userService = userService;
+        this.userRepository = userRepository;
     }
 
     public boolean createThumbUp(WebUser user, long reviewId) {
@@ -64,12 +68,12 @@ public class ThumbUpServiceImpl implements ThumbUpService{
         return result;
     }
 
-    public void deleteThumbUp(WebUser user, long reviewId) {
+    public void deleteThumbUp(long thumbUpId, long userId) {
 
-        Review review = reviewRepository.findById(reviewId).orElseThrow();
-        ThumbUp thumbUp = thumbUpRepository.findByWebUserAndReview(user, review).orElseThrow();
-
-        thumbUpRepository.delete(thumbUp);
+        ThumbUp thumbUp = thumbUpRepository.findById(thumbUpId).orElseThrow(() -> new BusinessLogicException(ExceptionCode.ID_NOT_FOUND));
+        if(userId == thumbUp.getWebUser().getUserId()){
+            thumbUpRepository.delete(thumbUp);
+        } else {new BusinessLogicException(ExceptionCode.USER_IS_NOT_MATCH);}
     }
 
 //  좋아요 중복 체크
